ml-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
numpy==1.26.2
scikit-learn==1.4.0
pandas==2.1.4
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0

